requirements.txt
numpy~=1.23.4
openai~=0.25.0
tenacity~=8.1.0
python-dotenv~=0.21.0
pandas~=1.5.1
tqdm~=4.64.1
matplotlib~=3.6.2
seaborn~=0.12.1
langchain~=0.0.73
retry~=0.9.2
setuptools~=60.2.0
lllm~=0.0.1
gensim~=4.3.0